compute multipole in the center of mass
